==> 02_rScripts/01_IdentifyingStands.Rmd <==
---
title: "Nunery Keeton Revisited"
author: "Nikolaus Bates-Haus"
output:
  pdf_document: default
  html_document: default
editor_options:
  markdown:
    wrap: 72
---

```{r setup, include=FALSE}
# Pre-load libraries that issue warnings on startup so these
# warnings don't corrupt the text later.
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(reshape2)) # for melt()
suppressPackageStartupMessages(library(RSQLite))
suppressPackageStartupMessages(library(htmltools))
suppressPackageStartupMessages(library(rmarkdown))
suppressPackageStartupMessages(library(dbplyr))
```

# Identifying Stands

In [@nunery2010, Table 1], 32 plots are listed, with the first column,
"FIA plot code", identifying each plot. As described in [@shaw2019] and
[@shaw2020], the method used to compute FVS stand identifiers
subsequently changed, so these codes no longer align with the stand
identifiers in the FVS tables in the published FIA data.

```{r}
nk_table1 <- read_csv(
  "../00_rawData/NK_Table_1.csv",
  col_types = cols(`FIA plot code` = col_character())
)
knitr::kable(nk_table1, caption = 'Nunery Keeton Table 1')
```

To re-align NK's plot codes with the FIA published FVS stand
identifiers, we note that plot codes are a concatenation of other
fields, as described in [@shaw2020, Appendix A], Column list and
translation logic for records in FVS StandInit and PlotInit tables:

> StandInit_cond table: StandInit_cond.Stand_ID = concatenation of:
> STATECD(4) + INVYR(4) + CYCLE(2) + SUBCYCLE(2) + UNITCD(2) +
> COUNTYCD(3) + PLOT(5) + CONDID(1).
>
> StandInit_plot table: StandInit_plot.Stand_ID = concatenation of:
> PLOT.STATECD(4) + PLOT.INVYR(4) + PLOT.CYCLE(2) + PLOT.SUBCYCLE(2) +
> PLOT.UNITCD(2) + PLOT.COUNTYCD(3) + PLOT.PLOT(5).
>
> PlotInit_plot table: PlotInit_plot.Stand_ID = concatenation of:
> PLOT.STATECD(4) + PLOT.INVYR(4) + PLOT.CYCLE(2) + PLOT.SUBCYCLE(2) +
> PLOT.UNITCD(2) + PLOT.COUNTYCD(3) + PLOT.PLOT(5).

Earlier methods of computing Stand_ID from FIA data used a 2-character
STATECD, and omitted CYCLE and SUBCYCLE. Using this information, we can
parse STATECD, INVYR, UNITCD, COUNTYCD, and PLOT from the
`FIA plot code` in NK.

```{r}
nk_table1_expanded <- nk_table1 |>
  mutate(STATECD = substr(`FIA plot code`, 1, 2)) |>
  mutate(INVYR = substr(`FIA plot code`, 3, 6)) |>
  mutate(UNITCD = substr(`FIA plot code`, 7, 8)) |>
  mutate(COUNTYCD = substr(`FIA plot code`, 9, 11)) |>
  mutate(PLOT = substr(`FIA plot code`, 12, 16))
nk_plot_codes <- nk_table1_expanded |>
  select(`FIA plot code`, STATECD, INVYR, UNITCD, COUNTYCD, PLOT)
knitr::kable(nk_plot_codes)
```

In addition, the way INVYR is assigned has changed, so the NK INVYR no
longer aligns with the inventory years in the FIA data for 27 of the 32
plots.

```{r}
fia = DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')
fia_cond = tbl(fia, 'COND')
# Note: we're joining tables across data providers so we have to copy one to the other.
# We do a right join so that we can copy the small table (nk_plot_codes) to
# the SQLite engine, where it can perform the join, while using the small table
# to select the output rows.
# We rename INVYR in NK to NK_INVYR, and in FIA to FIA_INVYR, so that we
# can compare the two.
matching_cond <- fia_cond |>
  rename(FIA_INVYR=INVYR) |>
  right_join(
    nk_plot_codes |> rename(NK_INVYR=INVYR),
    by=join_by(STATECD, UNITCD, COUNTYCD, PLOT),
    copy=TRUE
  ) |>
  collect()

DBI::dbDisconnect(fia)

# Group by STATECD, UNITCD, COUNTYCD and PLOT to see how NK INVYR does and
# doesn't line up with FIA INVYR
cond_grp <- matching_cond |>
  mutate(INVYR_MATCHES=ifelse(NK_INVYR==FIA_INVYR, 1, 0)) |>
  group_by(STATECD, UNITCD, COUNTYCD, PLOT) |>
  summarize(
    NUM_INVYR=n(),
    NUM_INVYR_MATCHES=sum(INVYR_MATCHES),
    .groups = "keep"
  )
knitr::kable(cond_grp)
```

To find corresponding Stand_IDs where INVYR does not align between the
two data sets, we note that NK's "Starting stand age" matches
`FIA.COND.STDAGE`, uniquely identifying all but one of the stands used
in NK.

```{r}
nk_plots_with_age <- nk_table1_expanded |>
  select(`FIA plot code`, STATECD, INVYR, UNITCD, COUNTYCD, PLOT,
         `Starting stand age`, `Slope (%)`, `Aspect (degrees)`,
         `Basal area (m2/ha)`) |>
  rename(NK_INVYR=INVYR)
fia_cond_with_age <- matching_cond |>
  select(STATECD, FIA_INVYR, CYCLE, SUBCYCLE, CONDID, UNITCD, COUNTYCD, PLOT,
         STDAGE, FLDAGE, SLOPE, ASPECT, BALIVE) |>
  left_join(
    nk_plots_with_age, by=join_by(STATECD, UNITCD, COUNTYCD, PLOT)
  )
unique_cond <- fia_cond_with_age |>
  filter(
    NK_INVYR==FIA_INVYR | (
      NK_INVYR!=FIA_INVYR & `Starting stand age`==STDAGE
    ),
    FIA_INVYR<=2005
  )

nk_to_fia <- unique_cond |>
  select(`FIA plot code`, NK_INVYR, 
         STATECD, FIA_INVYR, CYCLE, SUBCYCLE, CONDID, UNITCD, COUNTYCD, PLOT) |>
  rename(INVYR = FIA_INVYR)
write_csv(nk_to_fia, "../03_outputs/nk_to_fia.csv")

knitr::kable(unique_cond)
```

Matching on stand age results in a uniquely identified FIA condition for
each NK plot except NK plot 2320030901702963, for which both inventory
years 2003 and 2008 match the stand age. Because all other plots match
stands in inventory years from 1999 to 2005, we restrict to inventory
years \<= 2005, which results in selecting the 2003 inventory year for
this condition. We observe that slope, aspect, and BALIVE metrics for
the selected conditions match those for the corresponding NK plots,
which confirms the matches.

From these conditions we are able to construct updated Stand_IDs for use
with the FIA tables pre-built for FVS.

Note that the Stand IDs visible in the FVS UI are different from these;
the FVS UI Stand ID consists of STATECD(2) + INVYR(2) + COUNTYCD(3) +
PLOT(5), and if using FVS_STANDINIT_COND, then also + CONDID(1).

```{r}
translated_stands <- unique_cond |>
  mutate(
    STATECD  = as.numeric(STATECD),
    UNITCD   = as.numeric(UNITCD),
    COUNTYCD = as.numeric(COUNTYCD),
    PLOT     = as.numeric(PLOT)
  ) |>
  mutate(FVS_STAND_ID=sprintf(paste0(
    '%02d',  '%02d',           '%03d',   '%05d'),
    STATECD, FIA_INVYR %% 100, COUNTYCD, PLOT
  )) |>
  mutate(STAND_ID_PLOT=sprintf(paste0(
    '%04d' , '%04d'    , '%02d', '%02d'  , '%02d', '%03d'  , '%05d'),
    STATECD, FIA_INVYR , CYCLE , SUBCYCLE, UNITCD, COUNTYCD, PLOT
  )) |>
  mutate(STAND_ID_COND=paste0(STAND_ID_PLOT, CONDID)) |>
  select(`FIA plot code`, FVS_STAND_ID, STAND_ID_PLOT, STAND_ID_COND)
knitr::kable(translated_stands)
```

We can then fetch those IDs from the FIA-provided FVS tables to
determine the number of matching rows from each corresponding table,
confirming that each updated Stand ID corresponds to a unique stand in
the provided FVS tables.

```{r}
fia = DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')

fia_fvs_plotinit_plot = tbl(fia, 'FVS_PLOTINIT_PLOT')
matching_plotinit_plot <- fia_fvs_plotinit_plot |>
  right_join(translated_stands |> rename(STAND_ID=STAND_ID_PLOT),
    by=join_by(STAND_ID), copy=TRUE) |>
  collect()
matching_plotinit_plot_grp <- matching_plotinit_plot |>
  group_by(STAND_ID) |> summarize(NUM_PLOTS=n(), .groups = "keep") |>
  rename(FVS_PLOTINIT_PLOT=STAND_ID)

fia_fvs_standinit_cond = tbl(fia, 'FVS_STANDINIT_COND')
matching_standinit_cond <- fia_fvs_standinit_cond |>
  right_join(translated_stands |> rename(STAND_ID=STAND_ID_COND),
    by=join_by(STAND_ID), copy=TRUE) |>
  collect()
matching_standinit_cond_grp <- matching_standinit_cond |>
  group_by(STAND_ID) |> summarize(NUM_CONDS=n(), .groups = "keep") |>
  rename(FVS_STANDINIT_COND=STAND_ID)

fia_fvs_standinit_plot = tbl(fia, 'FVS_STANDINIT_PLOT')
matching_standinit_plot <- fia_fvs_standinit_plot |>
  right_join(translated_stands |> rename(STAND_ID=STAND_ID_PLOT),
    by=join_by(STAND_ID), copy=TRUE) |>
  collect()

DBI::dbDisconnect(fia)

matching_standinit_plot_grp <- matching_standinit_plot |>
  group_by(STAND_ID) |> summarize(NUM_COND_PLOTS=n(), .groups = "keep") |>
  rename(FVS_STANDINIT_PLOT=STAND_ID)

nk_to_fvs <- translated_stands |>
  left_join(matching_plotinit_plot_grp, by=join_by(STAND_ID_PLOT==FVS_PLOTINIT_PLOT)) |>
  left_join(matching_standinit_cond_grp, by=join_by(STAND_ID_COND==FVS_STANDINIT_COND)) |>
  left_join(matching_standinit_plot_grp, by=join_by(STAND_ID_PLOT==FVS_STANDINIT_PLOT))

write_csv(nk_to_fvs, "../03_outputs/nk_to_fvs.csv")

knitr::kable(nk_to_fvs)
```

Note that FIA provides three different means of creating an FVS Stand
from FIA data (descriptions from the FVS User Interface for the
respective tables):

-   FVS_StandInit_Plot: Inventory Plots
-   FVS_PlotInit_Plot: Inventory Subplots
-   FVS_StandInit_Cond: Conditions

Next we need to determine which of these, if any, corresponds to the
input used in NK.
