==> 02_rScripts/03_NKNoManagement.Rmd <==
---
title: "Nunery Keeton Revisited"
author: "Nikolaus Bates-Haus"
output:
  pdf_document: default
  html_document: default
editor_options:
  markdown:
    wrap: 72
---

```{r setup, include=FALSE}
# Pre-load libraries that issue warnings on startup so these
# warnings don't corrupt the text later.
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(reshape2)) # for melt()
suppressPackageStartupMessages(library(RSQLite))
suppressPackageStartupMessages(library(htmltools))
suppressPackageStartupMessages(library(rmarkdown))
suppressPackageStartupMessages(library(dbplyr))
suppressPackageStartupMessages(library(measurements))
```

# Reproducing the NoManagement Scenario

## Configuring FVS

FVS Software with `Release date: 20240401` was used.

FIA data version:

```{r}
fia <- DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')
fia_ref_fiadb_version <- tbl(fia, 'REF_FIADB_VERSION')
version <- fia_ref_fiadb_version |>
  filter(INSTALL_TYPE == 'RELEASE') |>
  arrange(desc(CREATED_DATE)) |>
  head(1) |>
  collect()
DBI::dbDisconnect(fia)
knitr::kable(version)
```

### Stands

Stands can be found in the FVS UI using FVS_STAND_ID from the table
above.

-   Inventory Data Tables = FVS_StandInit_Plot
-   Variants = ne: Northeast
-   Groups: All_FIA_Plots
-   Find stand(s): = 239901700110

Clicking `Find` will then filter to the single matching stand, which can
be added to the run.

Repeat for each of the above stands.

### Time

FVS will not run prior to the inventory year; it's unclear what NK did
to get their chart to go back prior to 2005. We will instead use 2005 as
the common starting year.

NK ends in 2164; we adjust this to 2165 so that all cycles are 10 years
long. Note that FVS does not include the ending year unless it is at the
start of a new cycle, so we set the common ending year to 2166.

All events in NK take place on 10 year intervals, therefore we select 10
years as the reporting interval.

### Regeneration

NK regeneration rates are in NK Table 4, in seedlings per hectare.

In addition to scenario-specific regeneration, NK 2.4 states,
"Background regeneration rates (intermediate to shade tolerant species
only), input at 10 year intervals, emulated natural regeneration within
stands, independent of forest management activities."

This background regeneration rate is in NK Table 4.

```{r}
table4 <- read_csv(
  "../00_rawData/NK_Table_4.csv",
  col_types = cols(
    `Management scenario` = col_character(),
    .default = col_number(),
  )
)

# Table4 uses scientific names; translate to get FVS Species Code
fvs_species_codes <- read_csv(
    "../01_processedInputData/FVSne_Overview_Table_3.2.1.csv",
    col_types = "iiciccc"
  ) |>
  select("Scientific Name", "Species Code", "Common Name")

# table4 has one column per species. We wish to add observations per species
# with different types, so pivot the table to have one row per species.
table4_rot <- table4 |>
  pivot_longer(cols = !`Management scenario`) |>
  pivot_wider(names_from = `Management scenario`) |>
  rename(`Scientific Name`=name) |>
  left_join(fvs_species_codes, by=join_by(`Scientific Name`))

# Values in NK are seedlings per hectare; FVS needs seedlings per acre
table4_rot$Clearcut <-
  round(conv_unit(table4_rot$Clearcut, 'hectare', 'acre'))
table4_rot$Shelterwood <-
  round(conv_unit(table4_rot$Shelterwood, 'hectare', 'acre'))
table4_rot$`ITS_Low Retention` <-
  round(conv_unit(table4_rot$`ITS_Low Retention`, 'hectare', 'acre'))
table4_rot$`ITS_High Retention` <-
  round(conv_unit(table4_rot$`ITS_High Retention`, 'hectare', 'acre'))
table4_rot$Background <-
  round(conv_unit(table4_rot$Background, 'hectare', 'acre'))

knitr::kable(
  table4_rot |>
  select(`Scientific Name`, Background, `Species Code`, `Common Name`) |>
  arrange(desc(Background))
)
```

In FVS, go to Simulate -\> Components -\> Management From Categories,
select "Planting & Natural Regeneration" From Components, select
"Plant/Natural with Partial Estab Model"

- Component title: "Baseline Regen"
- Schedule the date of disturbance: Schedule by condition
- Create a condition: Every cycle, every even cycle, ...
- Condition title: Every cycle
- Years before condition can become true again: 0
- The modulus of cycle number: 1 = every cycle
- Number of years after condition is found true: 0
- Sprouting: On

Note that this leaves Percent survival at 100, and average age and
height empty.

In freeform, this looks like:

```         
Estab              0
Sprout
Natural            0        SM      1221      100.                             0
Natural            0        AB       610      100.                             0
Natural            0        EH       153      100.                             0
Natural            0        RS       153      100.                             0
Natural            0        YB       153      100.                             0
Natural            0        RM       153      100.                             0
```

Switch to freeform and add all the rows, fix the titles.

Select `Save in run` to apply to `Grp: All_FIA_Plots`.

### Carbon

To set the carbon calculations to be metric (Metric tons of carbon per
hectare), go to Simulate -\> Components -\> Keywords, and select:

-   Extensions: Fire and Fuels Extension
-   Keywords: CarbCalc: Set the carbon accounting parameters.
-   Component title: CarbCalc: Metric
-   Biomass predictions: 0 = Use FFE biomass predictions
-   Units: 1 = Metric (metric tons carbon/hectare)
-   Note: Annual root decay rate (proportion per year) remains at its
    default value of 0.0425.

Select `Save in run` to apply to `Grp: All_FIA_Plots`.

## Run FVS

The main output will be Carbon and fuels. Stand visualization, Tree
lists, and Inventory statistics may also be interesting.


## Carbon Projection

### By Plot

Read FVS output for carbon storage by plot. FVS column names are a bit
idiosyncratic, so clean those up.

NK states that Fig 2 sums carbon from aboveground live, standing dead,
and down dead. We plot this, as well as just aboveground live and stand total
carbon for comparison.

```{r}
NoManagement_Carbon_ByPlot <- read_csv(
    "../03_outputs/NoManagement_Carbon_ByPlot.csv",
    col_types = cols(
      StandID = col_character()
    )
  ) |>
  rename(Aboveground_Live = Aboveground_Total_Live) |>
  rename(Down_Dead = Forest_Down_Dead_Wood) |>
  mutate(Aboveground_Dead = Standing_Dead + Down_Dead) |>
  mutate(Aboveground_Carbon = Aboveground_Live + Aboveground_Dead) |>
  select(StandID, Year, Total_Stand_Carbon, Aboveground_Carbon,
         Aboveground_Live, Aboveground_Dead, Standing_Dead, Down_Dead)
knitr::kable(NoManagement_Carbon_ByPlot)
```

Plot these against NK Fig 2 "NoManagement" scenario. NK Fig 2 has three
different scales on the horizontal axis:

1.  From 1995 to 2005, major ticks are every 5 years
2.  From 2005 to 2155, major ticks are every 10 years
3.  From 2155 to 2164, major ticks are every 9 years

NK states that data prior to 2005 is projected from mean growth rate. To
simplify the chart, we omit data prior to 2005, place major tick marks
every 10 years, and nudge the end date from 2164 to 2165.

Note that NK Fig2 year 2164 shows an anomalous reduction in carbon
storage in the NoManagement scenario; it might be the right thing to do
to omit 2164 / 2165 as well.

```{r}
fig2 <- read_csv(
  "../03_outputs/nk_fig2_reconstructed.csv",
  col_types = cols(
    Year = col_integer(),
    NoManagement = col_double(),
    ClearcutHigh = col_double(),
    ClearcutLow = col_double(),
    ShelterwoodHigh = col_double(),
    ShelterwoodLow = col_double(),
    ITS_LowHigh = col_double(),
    ITS_LowLow = col_double(),
    ITS_HighHigh = col_double(),
    ITS_HighLow = col_double()
  )
)

NoManagement_Carbon_ByPlot_ByYear <- NoManagement_Carbon_ByPlot |>
  group_by(Year) |>
  summarize(
    Total_Carbon=mean(Total_Stand_Carbon),
    Aboveground_Carbon=mean(Aboveground_Carbon),
    Aboveground_Live=mean(Aboveground_Live),
    .groups = "keep"
  ) |>
  filter(Year >= 2005) |>
  # Join in the NoManagement scenario from fig2 for comparison
  full_join(fig2 |> select(Year,NoManagement), by=join_by(Year)) |>
  rename(NK_Fig2_NoManagement = NoManagement)

write_csv(NoManagement_Carbon_ByPlot_ByYear, "../03_outputs/NoManagement_Carbon_ByPlot_ByYear.csv")

ggplot(
    data = melt(NoManagement_Carbon_ByPlot_ByYear, id.vars = "Year"),
    mapping = aes(x = Year, y = value, color = variable, linetype = variable)
  ) +
  ggtitle("FVS Carbon Projection for NK Plots") +
  ylab("Carbon (Mg/ha)") +
#  theme(legend.title = element_blank()) +
  theme(legend.title = element_blank(), legend.position = "bottom") +
  scale_linetype_manual(values = c("solid", "solid", "solid", "dashed")) +
  geom_line() +
  coord_cartesian(xlim = c(2005, 2165), ylim = c(0, 320)) +
  scale_x_continuous(breaks=seq(2005,2165,20)) +
  scale_y_continuous(breaks=seq(0,320,20))
```

From the chart it seems evident that aboveground live carbon is the best match
for the NoManagement scenario in NK Fig 2. We confirm this by computing
RMS error for the residuals between our FVS runs and the NK Fig 2 data:

```{r}
Plots_RMSE <- data.frame(
  Total_Carbon = sqrt(mean(
    (NoManagement_Carbon_ByPlot_ByYear$Total_Carbon - NoManagement_Carbon_ByPlot_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Carbon = sqrt(mean(
    (NoManagement_Carbon_ByPlot_ByYear$Aboveground_Carbon - NoManagement_Carbon_ByPlot_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Live = sqrt(mean(
    (NoManagement_Carbon_ByPlot_ByYear$Aboveground_Live -
       NoManagement_Carbon_ByPlot_ByYear$NK_Fig2_NoManagement)^2
  ))
)
ggplot(
    data = melt(Plots_RMSE, id = NULL, variable.name = "Series", value.name = "RMSE"),
    mapping = aes(x = Series, y = RMSE)
  ) +
  ggtitle("FVS Carbon Projection by FIA Plot") +
  geom_col()
```

Note that NK describe using Aboveground Carbon, including both live and dead,
for their chart, but the best fit is from Aboveground Live, with RMSE \~= 12.

```{r}
Plots_RMSE |> select(Aboveground_Live)
```

### By Subplot

To confirm that NK used the FVS_StandInit_Plot table, we examine the
RMSE between projections from other tables and the NK Fig 2 values.

We repeat the above FVS run, but select table FVS_PlotInit_Plot and
select all subplots using the translated stand IDs. We apply the same
configuration to all stands as we did for FVS_StandInit_Plot. We load
and clean the results in the same manner.

```{r}
NoManagement_Carbon_BySubplot_ByYear <- read_csv(
    "../03_outputs/NoManagement_Carbon_BySubplot.csv",
    col_types = cols(StandID = col_character())
  ) |>
  rename(Total_Carbon = Total_Stand_Carbon) |>
  rename(Aboveground_Live = Aboveground_Total_Live) |>
  rename(Down_Dead = Forest_Down_Dead_Wood) |>
  mutate(Aboveground_Dead = Standing_Dead + Down_Dead) |>
  mutate(Aboveground_Carbon = Aboveground_Live + Aboveground_Dead) |>
  select(StandID, Year, Total_Carbon, Aboveground_Carbon,
         Aboveground_Live, Aboveground_Dead, Standing_Dead, Down_Dead) |>
  group_by(Year) |>
  summarize(
    Total_Carbon=mean(Total_Carbon),
    Aboveground_Carbon=mean(Aboveground_Carbon),
    Aboveground_Live=mean(Aboveground_Live),
    .groups = "keep"
  ) |>
  filter(Year >= 2005) |>
  full_join(fig2 |> select(Year,NoManagement), by=join_by(Year)) |>
  rename(NK_Fig2_NoManagement = NoManagement)

ggplot(
    data = melt(NoManagement_Carbon_BySubplot_ByYear, id.vars = "Year"),
    mapping = aes(x = Year, y = value, color = variable)
  ) +
  ggtitle("FVS Carbon Projection by FIA Subplot") +
  ylab("Carbon (Mg/ha)") +
  theme(legend.title = element_blank()) +
  geom_line() +
  coord_cartesian(xlim = c(2005, 2165), ylim = c(0, 320)) +
  scale_x_continuous(breaks=seq(2005,2165,20)) +
  scale_y_continuous(breaks=seq(0,320,20))
```

Examine RMSE for the residuals:

```{r}
Subplots_RMSE <- data.frame(
  Total_Carbon = sqrt(mean(
    (NoManagement_Carbon_BySubplot_ByYear$Total_Carbon - NoManagement_Carbon_BySubplot_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Carbon = sqrt(mean(
    (NoManagement_Carbon_BySubplot_ByYear$Aboveground_Carbon - NoManagement_Carbon_BySubplot_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Live = sqrt(mean(
    (NoManagement_Carbon_BySubplot_ByYear$Aboveground_Live - NoManagement_Carbon_BySubplot_ByYear$NK_Fig2_NoManagement)^2
  ))
)
ggplot(
    data = melt(Subplots_RMSE, id = NULL, variable.name = "Series", value.name = "RMSE"),
    mapping = aes(x = Series, y = RMSE)
  ) +
  ggtitle("FVS Carbon Projection by FIA Subplot") +
  geom_col()
```

Note that all RMSE for the subplot projections are significantly worse
than for the plot projections.

### By Condition

We repeat the exercise using the FVS_StandInit_Cond table, which creates
FVS stands for corresponding FIA conditions.

We select all conditions using the translated stand IDs. Note that plot
360304303966 has 3 conditions; we include all three. We apply the same
configuration to all stands as we did for FVS_StandInit_Plot. We load
and clean the results in the same manner.

```{r}
NoManagement_Carbon_ByCondition_ByYear <- read_csv(
    "../03_outputs/NoManagement_Carbon_ByCondition.csv",
    col_types = cols(StandID = col_character())
  ) |>
  rename(Total_Carbon = Total_Stand_Carbon) |>
  rename(Aboveground_Live = Aboveground_Total_Live) |>
  rename(Down_Dead = Forest_Down_Dead_Wood) |>
  mutate(Aboveground_Dead = Standing_Dead + Down_Dead) |>
  mutate(Aboveground_Carbon = Aboveground_Live + Aboveground_Dead) |>
  select(StandID, Year, Total_Carbon, Aboveground_Carbon,
         Aboveground_Live, Aboveground_Dead, Standing_Dead, Down_Dead) |>
  group_by(Year) |>
  summarize(
    Total_Carbon=mean(Total_Carbon),
    Aboveground_Carbon=mean(Aboveground_Carbon),
    Aboveground_Live=mean(Aboveground_Live),
    .groups = "keep"
  ) |>
  filter(Year >= 2005) |>
  full_join(fig2 |> select(Year,NoManagement), by=join_by(Year)) |>
  rename(NK_Fig2_NoManagement = NoManagement)

ggplot(
    data = melt(NoManagement_Carbon_ByCondition_ByYear, id.vars = "Year"),
    mapping = aes(x = Year, y = value, color = variable)
  ) +
  ggtitle("FVS Carbon Projection by FIA Condition") +
  ylab("Carbon (Mg/ha)") +
  theme(legend.title = element_blank()) +
  geom_line() +
  coord_cartesian(xlim = c(2005, 2165), ylim = c(0, 320)) +
  scale_x_continuous(breaks=seq(2005,2165,20)) +
  scale_y_continuous(breaks=seq(0,320,20))
```

Examine RMSE for the residuals:

```{r}
Conditions_RMSE <- data.frame(
  Total_Carbon = sqrt(mean(
    (NoManagement_Carbon_ByCondition_ByYear$Total_Carbon - NoManagement_Carbon_ByCondition_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Carbon = sqrt(mean(
    (NoManagement_Carbon_ByCondition_ByYear$Aboveground_Carbon - NoManagement_Carbon_ByCondition_ByYear$NK_Fig2_NoManagement)^2
  )),
  Aboveground_Live = sqrt(mean(
    (NoManagement_Carbon_ByCondition_ByYear$Aboveground_Live - NoManagement_Carbon_ByCondition_ByYear$NK_Fig2_NoManagement)^2
  ))
)
ggplot(
    data = melt(Conditions_RMSE, id = NULL, variable.name = "Series", value.name = "RMSE"),
    mapping = aes(x = Series, y = RMSE)
  ) +
  ggtitle("FVS Carbon Projection by FIA Condition") +
  geom_col()
```

RMSE for Aboveground Live by Condition is \~13, which is close to that
for Aboveground Live by Plot.

```{r}
Conditions_RMSE |> select(Aboveground_Live)
```
