==> 02_rScripts/04_CarbonCalc.Rmd <==
---
title: "Nunery Keeton Revisited"
author: "Nikolaus Bates-Haus"
output:
  pdf_document: default
  html_document: default
editor_options:
  markdown:
    wrap: 72
---

```{r setup, include=FALSE}
# Pre-load libraries that issue warnings on startup so these
# warnings don't corrupt the text later.
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(reshape2)) # for melt()
suppressPackageStartupMessages(library(RSQLite))
suppressPackageStartupMessages(library(htmltools))
suppressPackageStartupMessages(library(rmarkdown))
suppressPackageStartupMessages(library(dbplyr))
suppressPackageStartupMessages(library(measurements))
```

# Carbon Calculation

[@Herbert2023] use FIA measurement of biomass to determine carbon. FIA
also provides above-groud carbon in the TREE table, in CARBON_AG. This
metric is, of course, not available on plots that FVS grows. For
consistent carbon calculation on both measured and grown plots, we can
use the FVS Fire and Fuels extension (FFE-FVS), which contains a Carbon
Submodel, to compute carbon. We compare FFE carbon calculation with
Jenkins and FIA carbon calculation at the end of this section.

## Stand Inventories

FIA has two notions of when a plot is inventoried:

*  Inventory year (INVYR): the nominal year for which an inventory is
   conducted. This is the "official" year, and is used to identify observations.

*  Measure Year (MEASYEAR): the actual year in which a plot was measured.
   For anything sensitive to real-world time, such as growth, this is the
   time measurement that matters. Note that MEASYEAR may be less than, equal
   to, or greater than INVYR.

The FIA PLOT table lists when plots were measured; the FIA COND table
lists observations from those measurements. We can use the above
established connection between FVS identifiers and FIA identifiers to
find subsequent measurement for the plots and conditions included in NK.

Create a single crosswalk between NK, FIA and FVS identifiers.

```{r}
nk_to_fia <- read_csv(
    "../03_outputs/nk_to_fia.csv",
    col_types = cols(
      `FIA plot code` = col_character(),
      .default = col_integer()
    )
  ) |>
  select(`FIA plot code`, STATECD, COUNTYCD, PLOT, CONDID)

nk_to_fvs <- read_csv(
    "../03_outputs/nk_to_fvs.csv",
    col_types = cols(
      `FIA plot code` = col_character(),
      FVS_STAND_ID = col_character(),
      STAND_ID_PLOT = col_character(),
      STAND_ID_COND = col_character(),
      .default = col_integer()
    )
  ) |>
  select(`FIA plot code`, FVS_STAND_ID)

stand_id_crosswalk <- nk_to_fia |>
  full_join(nk_to_fvs, by=join_by(`FIA plot code`)) |>
  # STAND_ID is state, county, plot - like FVS_STAND_ID with no year
  mutate(STAND_ID = sprintf(paste0(
    '%02d',  '%03d',   '%05d'),
    STATECD, COUNTYCD, PLOT
  ))

remove(nk_to_fia)
remove(nk_to_fvs)
```

Given the NK plot IDs, we can find the matching FIA plots.

```{r}
fia = DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')
fia_plot = tbl(fia, 'PLOT')

matching_plot <- fia_plot |>
  filter(INVYR >= 1999) |>
  semi_join(
    stand_id_crosswalk,
    by=join_by(STATECD, COUNTYCD, PLOT),
    copy=TRUE
  ) |>
  collect()

DBI::dbDisconnect(fia)

remove(fia, fia_plot)
```

### Remeasurement Timeline

To aid visual analysis, we create arbitrary ordinals for plots;
this will allow us to chart each plot in a single
horizontal band, against time, to create a visual timeline of plot
measurement.

To create these ordinals, we sort plots by:

1.  The first year measured, earliest first; then
2.  Break ties by the second year measured, earliest year first; then
3.  Break ties by the third year measured, earliest year first; then
4.  Break ties by the fourth year measured, earliest year first
5.  Assign stand number as dense rank row number.

This will put plots surveyed earlier towards the bottom of the chart,
creating an "upward and to the right" visual that is generally easy to read.

```{r}
year1 <- matching_plot |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(YEAR1 = min(MEASYEAR), .groups = "keep") |>
  ungroup() |>
  right_join(matching_plot, by=join_by(STATECD, COUNTYCD, PLOT))

year2 <- year1 |>
  filter(MEASYEAR > YEAR1) |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(YEAR2 = min(MEASYEAR), .groups = "keep") |>
  ungroup() |>
  right_join(year1, by=join_by(STATECD, COUNTYCD, PLOT))
  
year3 <- year2 |>
  filter(MEASYEAR > YEAR2) |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(YEAR3 = min(MEASYEAR), .groups = "keep") |>
  ungroup() |>
  right_join(year2, by=join_by(STATECD, COUNTYCD, PLOT))

year4 <- year3 |>
  filter(MEASYEAR > YEAR3) |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(YEAR4 = min(MEASYEAR), .groups = "keep") |>
  ungroup() |>
  right_join(year3, by=join_by(STATECD, COUNTYCD, PLOT))

# Compute stand numbers as ordinals according to inventory years
stand_num <- year4 |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarise(
    YEAR1 = min(YEAR1),
    YEAR2 = min(YEAR2),
    YEAR3 = min(YEAR3),
    YEAR4 = min(YEAR4),
    .groups = "keep"
  ) |>
  ungroup() |>
  arrange(YEAR1, YEAR2, YEAR3, YEAR4) |>
  mutate(STAND_NUM = row_number()) |>
  select(STAND_NUM, STATECD, COUNTYCD, PLOT)

remove(year1, year2, year3, year4)

# Paste stand numbers back on to plots
stand_inventories <- matching_plot |>
  left_join(stand_num, by=join_by(STATECD, COUNTYCD, PLOT)) |>
  group_by(STAND_NUM, MEASYEAR) |>
  rename(Year = MEASYEAR)

ggplot(
    data = stand_inventories,
    mapping = aes(x = Year, y = STAND_NUM)
  ) +
  geom_point()
```

### Changing Conditions

Conditions are numbered according to where they appear in
the plot for a given inventory. As such, the same physical condition 
can receive a different number in different inventories. For example,
if the condition at the center of subplot 1 changes from inventory A to inventory B, then
the characteristics of the condition numbered 1 will be different between the two inventories, even if there are
portions of the plot that still match what used to be condition 1. To
assign stand ordinals actually correctly, we would need to align
conditions by characteristics such as azimuth, forest type and stand
age. Because there are only a few plots with multiple conditions, we simplify
this analysis by omitting multi-condition plots.

```{r}
fia = DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')
fia_cond = tbl(fia, 'COND')
fia_ref_forest_type = tbl(fia, 'REF_FOREST_TYPE') |>
  select(VALUE, MEANING)

matching_cond <- fia_cond |>
  filter(INVYR >= 1999) |>
  semi_join(
    stand_id_crosswalk,
    # Omitting CONDID in the join will not filter out any conditions
    by=join_by(STATECD, COUNTYCD, PLOT),
    copy=TRUE
  ) |>
  # Dereference forest type codes
  left_join(fia_ref_forest_type, by = join_by(FORTYPCD == VALUE)) |>
  rename(FORTYPE = MEANING) |>
  # Paste MEASYEAR on to COND, because everything in FVS works off MEASYEAR
  collect() |>
  left_join(
    matching_plot |> select(STATECD, COUNTYCD, PLOT, INVYR, MEASYEAR),
    by=join_by(STATECD, COUNTYCD, PLOT, INVYR)
  )

DBI::dbDisconnect(fia)

remove(fia, fia_cond, fia_ref_forest_type)

# We are interested in plots that have more than one condition in any
# inventory year. Therefore we group by plot, and count the number of
# distinct conditions across all inventory years.
# This will not flag cases when a plot with a single condition changes
# condition type, e.g. when a plot is harvested.
matching_cond_multi <- matching_cond |>
  select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR) |>
  mutate(MULTI_FLAG = CONDID > 1)

knitr::kable(matching_cond_multi |> filter(MULTI_FLAG))
```

### Skipped

Survey may be skipped for various reasons captured on the condition 
in COND_STATUS_CD and COND_NONSAMPLE_REASN_CD. Codes other than 1 indicate
that forest was not inventoried in the condition.

Status Code Descriptions are given in [@Burrill2024]:

-   1 Accessible forest land
-   2 Nonforest land
-   3 Noncensus water
-   4 Census water
-   5 Nonsampled, possibility of forest land

Nonsample Reason Code Descriptions are given in [@Burrill2024]:

-   01 Outside U.S. boundary
-   02 Denied access area
-   03 Hazardous situation
-   05 Lost data
-   06 Lost plot
-   07 Wrong location
-   08 Skipped visit
-   09 Dropped intensified plot
-   10 Other
-   11 Ocean

We want to remove plots that are skipped for any reason.

```{r}
# Were any stands not sampled during an inventory?
matching_cond_skipped <- matching_cond |>
  select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, COND_STATUS_CD, COND_NONSAMPLE_REASN_CD) |>
  mutate(SKIPPED_FLAG = COND_STATUS_CD > 1)

knitr::kable(matching_cond_skipped |> filter(SKIPPED_FLAG))
```

### Disturbance

Three condition fields, DSTRBCD1/2/3, capture Disturbance codes. These indicate
types of disturbance to the condition other than disturbances related to
harvest.

Code Descriptions are given in [@Burrill2024]:

-   0 No visible disturbance.
-   10 Insect damage.
-   11 Insect damage to understory vegetation.
-   12 Insect damage to trees, including seedlings and saplings.
-   20 Disease damage.
-   21 Disease damage to understory vegetation.
-   22 Disease damage to trees, including seedlings and saplings.
-   30 Fire damage (from crown and ground fire, either prescribed or
    natural).
-   31 Ground fire damage.
-   32 Crown fire damage.
-   40 Animal damage.
-   41 Beaver (includes flooding caused by beaver).
-   42 Porcupine.
-   43 Deer/ungulate.
-   44 Bear (core optional).
-   45 Rabbit (core optional).
-   46 Domestic animal/livestock (includes grazing).
-   50 Weather damage.
-   51 Ice.
-   52 Wind (includes hurricane, tornado).
-   53 Flooding (weather induced).
-   54 Drought.
-   60 Vegetation (suppression, competition, vines).
-   70 Unknown / not sure / other.
-   80 Human-induced damage - any significant threshold of human-caused
    damage not described in the DISTURBANCE codes or in the TREATMENT
    codes.
-   90 Geologic disturbances.
-   91 Landslide.
-   92 Avalanche track.
-   93 Volcanic blast zone.
-   94 Other geologic event.
-   95 Earth movement / avalanches.

```{r}
# Who had disturbance?
matching_cond_disturbed <- matching_cond |>
  select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, DSTRBCD1, DSTRBCD2, DSTRBCD3) |>
  mutate(DISTURBED_FLAG = DSTRBCD1 > 0 | DSTRBCD2 > 0 | DSTRBCD3 > 0)

knitr::kable(matching_cond_disturbed |> filter(DISTURBED_FLAG))
```

### Treatment

Harvest-related disturbance is captured in the three columns TRTCD1/2/3.
These capture treatment code describing disturbance related to harvest.

Code Descriptions are given in [@Burrill2024]:

-   00 No observable treatment.
-   10 Cutting
-   20 Site preparation
-   30 Artificial regeneration
-   40 Natural regeneration
-   50 Other silvicultural treatment

```{r}
# Who had treatment?
matching_cond_treated <- matching_cond |>
  select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, TRTCD1, TRTCD2, TRTCD3) |>
  mutate(TREATED_FLAG = TRTCD1 > 0 | TRTCD2 > 0 | TRTCD3 > 0)

knitr::kable(matching_cond_treated |> filter(TREATED_FLAG))
```


### Timeline

We can combine mult-condition, skipped, disturbed, and harvested statuses
to create a single timeline that shows plot status since the start of NK.

Note that the various status are per condition, and we are charting per plot. To
accommodate, we introduce horizontal offset for conditions so that the status of
individual conditions within a plot is visible in the chart.

```{r}
matching_cond_status <- matching_cond |>
  
  # Merge all the matching_cond_<status> tables
  left_join(
    matching_cond_multi |> select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, MULTI_FLAG),
    by = join_by(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR)
  ) |>
  left_join(
    matching_cond_skipped |> select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, SKIPPED_FLAG),
    by = join_by(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR)
  ) |>
  left_join(
    matching_cond_disturbed |> select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, DISTURBED_FLAG),
    by = join_by(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR)
  ) |>
  left_join(
    matching_cond_treated |> select(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR, TREATED_FLAG),
    by = join_by(STATECD, COUNTYCD, PLOT, CONDID, MEASYEAR)
  ) |>

  # Collect the various status indicators into a single STATUS column
  mutate(STATUS = NA) |>
  mutate(STATUS = if_else(MULTI_FLAG, "Multi-Condition", STATUS)) |>
  mutate(STATUS = if_else(SKIPPED_FLAG, "Skipped", STATUS)) |>
  mutate(STATUS = if_else(DISTURBED_FLAG, "Disturbed", STATUS)) |>
  mutate(STATUS = if_else(TREATED_FLAG, "Treated", STATUS)) |>
  mutate(STATUS = if_else(is.na(STATUS), "Grown", STATUS))

stand_inventory_status <- matching_cond_status |>
  
  # To facilitate visual display, nudge MEASYEAR by condition
  # max(CONDID) == 3, so nudge by 1/4 per CONDID.
  mutate(MEASYEAR = MEASYEAR + (CONDID - 1)/4) |>
  
  # Use the single STATUS column to create separate columns for each status,
  # where the measurement year is the value for the column
  mutate(GROWN = if_else(STATUS == "Grown", MEASYEAR, NA)) |>
  mutate(SKIPPED = if_else(STATUS == "Skipped", MEASYEAR, NA)) |>
  mutate(TREATED = if_else(STATUS == "Treated", MEASYEAR, NA)) |>
  mutate(DISTURBED = if_else(STATUS == "Disturbed", MEASYEAR, NA)) |>
  mutate(MULTI_COND = if_else(STATUS == "Multi-Condition", MEASYEAR, NA)) |>
  
  # Add stand numbers
  left_join(stand_num, by = join_by(STATECD, COUNTYCD, PLOT)) |>
  select(STAND_NUM, GROWN, TREATED, DISTURBED, SKIPPED, MULTI_COND) |>

  # Melt and arrange for ggplot
  melt(id.vars = "STAND_NUM", variable.name="Status", value.name = "Year") |>
  filter(!is.na(Year)) |>
  select(STAND_NUM, Year, Status) |>
  arrange(STAND_NUM, Year, Status)

ggplot(
    data = stand_inventory_status,
    mapping = aes(x = Year, y = STAND_NUM, shape = Status, color = Status)
  ) +
  ggtitle("Stand Inventory Timeline") +
  geom_point()
```
> TODO nik: Plot 23 is multi-condition, but none of the conditions is flagged
> as such. Figure that out and fix it.

### Grow-Only Stands

[@Herbert2023] says "Where stands were coded as having ≥25% of the trees
or sample area affected by insects, disease, weather damage, or geologic
disturbances damage between inventory visits, we coded a binary,
non-fire disturbance attribute to have available as an explanatory
factor but did not drop them from the analysis set."

There are a subset of plots that are grow-only for
all subsequent inventories. We would like to compare the carbon in the
FIA inventories for the grow-only plots with the carbon projected by FVS.

We start by identifying those plots that are grow in all conditions in all years.
We need all conditions in the plot to be grow-only for all inventory years, since
FIA and FVS may diverge in their understanding of condition composition
of plots over time.

```{r}
rhs <- matching_cond_status |>
  mutate(GROWN_FLAG = !MULTI_FLAG & !SKIPPED_FLAG & !DISTURBED_FLAG & !TREATED_FLAG) |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(
    GROWN_FLAG = all(GROWN_FLAG),
    .groups = "keep"
  ) |>
  ungroup() |>
  filter(GROWN_FLAG)

stand_grow_only <- matching_plot |>
  semi_join(rhs, by = join_by(STATECD, COUNTYCD, PLOT)) |>
  left_join(stand_num, by = join_by(STATECD, COUNTYCD, PLOT)) |>
  left_join(stand_id_crosswalk, by = join_by(STATECD, COUNTYCD, PLOT)) |>
  # Add STDAGE and FORTYPE, because they'll be useful later
  left_join(
    matching_cond |> 
      select(STATECD, COUNTYCD, PLOT, FORTYPE, STDAGE, MEASYEAR),
    by = join_by(STATECD, COUNTYCD, PLOT, MEASYEAR)
  )

# rhs is just for the RHS of the semi_join
rm(rhs)
  
ggplot(
    data = semi_join(stand_inventory_status, stand_grow_only, by = join_by(STAND_NUM)),
    mapping = aes(x = Year, y = STAND_NUM, shape = Status, color = Status)
  ) +
  ggtitle("Grow-only Stand Inventory Timeline") +
  geom_point() +
  coord_cartesian(xlim = c(1999, 2021), ylim = c(0, 31))
```

## Carbon Accounting

[@fvsstaff2021] Module 2, section Audit 1, describes how to perform a
zero-year run of FVS. This applies the mapping from FIA plot to FVS
stand, then grows the stand for zero years, which keeps the trees as
they are in the FIA plot. This does however produce all the FVS output
data structures for the stand, which enables application of e.g. the
FFE carbon model. We do not apply the regeneration adjustments, since we are
simply computing carbon from the FIA inventory data. We do this for each
individual grow-only plot in each measurement year, then chart the observed
vs. predicted carbon for these plots.

Note that plots change over time, and FIA's way of dealing with this is
to change the conditions over time. FVS does not change conditions as a
plot ages; it grows trees assuming that the forest type of the plot
remains unchanged. The result is that FVS grows the conditions as
inventoried at the start of the window, whereas FIA will adjust those
conditions over time. For both projected and measured carbon, we use all
conditions in the plot as inventoried, to compare carbon projection vs.
measurement for the entire plot land area.

### Running FVS

We need two outputs from FVS to compare:

1. Projected growth for all grow-only stands for the time window 2000 - 2024
   with a 1-year time step.
2. Carbon calculation for each stand for each year in which it was measured.

The first is accomplished with a single run of FVS with a 1-year timestep. Note
that this will have observations for each stand for years in which that stand
was not measured; we will filter those out later.

The second is accomplished by creating a 0-year FVS run for each year in which
a stand was measured, and running all stands measured in that year.

First, determine the years in which at least one plot was measured:
```{r}
measyears <- stand_grow_only |>
  group_by(MEASYEAR) |>
  summarize(.groups = "keep") |>
  ungroup()

knitr::kable(measyears)
```

Next, we want to run FVS for each of those years, including the stands that were measured
in each year. Since FVS stand ID includes the measure year, we need to calculate
the FVS_STAND_IDs to use for the run.

```{r}
fvs_stand_grow_only <- stand_grow_only |>
  # Compute FVS_STAND_ID to use the correct year
  mutate(FVS_STAND_ID=sprintf(paste0(
    '%02d' , '%02d'         , '%03d'  , '%05d'),
    STATECD, MEASYEAR %% 100, COUNTYCD, PLOT
  )) |>
  # Compute STAND_ID, which omits the year
  mutate(STAND_ID=sprintf(paste0(
    '%02d' , '%03d'  , '%05d'),
    STATECD, COUNTYCD, PLOT
  )) |>
  select(STATECD, COUNTYCD, PLOT, MEASYEAR, STDAGE, FORTYPE, FVS_STAND_ID, STAND_ID) |>
  arrange(MEASYEAR, STAND_ID)

knitr::kable(fvs_stand_grow_only)
```

### FFE-FFE

FFE-FVS is the Fire and Fuels extension for FVS. It provides two methods of
computing carbon: FFE and Jenkins. The default is FFE; we start by comparing
FFE Carbon for measured vs. projected stands.

This gives us a plot of measured carbon over time.

> TODO nik: Grow_ should be something to indicate that we're not growing,
> and instead just converting FIA data to FVS data.

```{r}
FVS_Carbon <- NULL

for (measyear in measyears)
  fname <- paste0("../03_outputs/Grow_", sprintf("%04d", measyear), "_FVS_Carbon.csv")
  FVS_Carbon <- bind_rows(
      FVS_Carbon,
      read_csv(fname, col_types = cols(StandID = col_character()))
    )


fvs_stand_measured_ffe <- FVS_Carbon |>
  mutate(STAND_ID = paste0(
     substring(StandID, 1, 2),
     substring(StandID, 5, 12)
  )) |>
  semi_join(fvs_stand_grow_only, by = join_by(STAND_ID)) |>
  mutate(Aboveground_Live = Aboveground_Total_Live) |>
  left_join(
    fvs_stand_grow_only |> select(STAND_ID, MEASYEAR, STDAGE, FORTYPE),
    by = join_by(STAND_ID, Year == MEASYEAR)
  ) |>
  select(STAND_ID, Year, Aboveground_Live, STDAGE, FORTYPE)

remove(FVS_Carbon, fname, measyear)

ggplot(
    data = fvs_stand_measured_ffe,
    mapping = aes(x = Year, y = Aboveground_Live, color = STAND_ID)
  ) +
  geom_point() +
  coord_cartesian(xlim = c(2000, 2025), ylim = c(0, 300)) +
  scale_x_continuous(breaks=seq(2000,2025,10)) +
  ggtitle("Grow-Only Stands\nFFE Measured Carbon (Mg / ha)")
```

We then run FVS on all grow-only stands from the start year with a one-year timestep to
get projected FFE carbon.

```{r}
fvs_stand_projected_ffe <- read_csv(
    "../03_outputs/NoManagement_Carbon_ByPlotYearly.csv",
    col_types = cols(StandID = col_character())
  ) |>
  left_join(stand_id_crosswalk, by = join_by(StandID == FVS_STAND_ID)) |>
  semi_join(fvs_stand_grow_only, by = join_by(STAND_ID)) |>
  rename(Aboveground_Live = Aboveground_Total_Live) |>
  select(STAND_ID, Year, Aboveground_Live)

ggplot(
    data = fvs_stand_projected_ffe,
    mapping = aes(x = Year, y = Aboveground_Live, color = STAND_ID)
  ) +
  geom_point() +
  coord_cartesian(xlim = c(2000, 2025), ylim = c(0, 300)) +
  scale_x_continuous(breaks=seq(2000,2025,10)) +
  ggtitle("Grow-Only Stands\nFFE Projected Carbon (Mg / ha)")
```

We can then plot the measured vs. projected carbon.

```{r}
stand_first_measured <- stand_inventories |>
  group_by(STATECD, COUNTYCD, PLOT) |>
  summarize(first_measure_year = min(Year), .groups = "keep") |>
  ungroup() |>
  left_join(stand_id_crosswalk, by = join_by(STATECD, COUNTYCD, PLOT))

compare_ffe <- fvs_stand_measured_ffe |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  rename(Measured = Aboveground_Live) |>
  left_join(
    fvs_stand_projected_ffe |> select(STAND_ID, Year, Aboveground_Live),
    by = join_by(STAND_ID, Year)
  ) |>
  rename(Projected = Aboveground_Live) |>
  left_join(stand_first_measured, by = join_by(STAND_ID)) |>
  mutate(`Projection Years` = Year - first_measure_year)

ggplot(
    data = compare_ffe
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = Measured, y = Projected, color = `Projection Years`, shape = `Forest Type`)
  ) +
  ggtitle("Grow-Only Stands\nFFE Measured vs. Projected Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

Same, but color by stand age

```{r}
ggplot(
    data = compare_ffe
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = Measured, y = Projected, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("Grow-Only Stands\nFFE Measured vs. Projected Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

### FFE-Jenkins

The second method of computing carbon provided by FFE-FVS is Jenkins. Compare
measured vs. projected carbon using FFE Jenkins.

To gather measured carbon using the Jenkins method, we must re-run FVS,
adding the Jenkins keyword for the fire and fuels extension.

For measured carbon, we re-run all the plots during the years they are measured.

```{r}
FVS_Carbon_Jenkins <- NULL

for (measyear in measyears)
  fname <- paste0(
    "../03_outputs/Grow_",
    sprintf("%04d", measyear),
    "_FVS_Carbon_Jenkins.csv"
  )
  FVS_Carbon_Jenkins <- bind_rows(
      FVS_Carbon_Jenkins,
      read_csv(fname, col_types = cols(StandID = col_character()))
    )

fvs_stand_measured_jenkins <- FVS_Carbon_Jenkins |>
  mutate(STAND_ID = paste0(
     substring(StandID, 1, 2),
     substring(StandID, 5, 12)
  )) |>
  semi_join(fvs_stand_grow_only, by = join_by(STAND_ID)) |>
  mutate(Aboveground_Live = Aboveground_Total_Live) |>
  left_join(
    fvs_stand_grow_only |> select(STAND_ID, MEASYEAR, STDAGE, FORTYPE),
    by = join_by(STAND_ID, Year == MEASYEAR)
  ) |>
  select(STAND_ID, Year, Aboveground_Live, STDAGE, FORTYPE)

remove(FVS_Carbon_Jenkins, fname, measyear)


ggplot(
    data = fvs_stand_measured_jenkins,
    mapping = aes(x = Year, y = Aboveground_Live, color = STAND_ID)
  ) +
  geom_point() +
  coord_cartesian(xlim = c(2000, 2025), ylim = c(0, 300)) +
  scale_x_continuous(breaks=seq(2000,2025,10)) +
  ggtitle("Grow-Only Stands\nJenkins Measured Carbon (Mg / ha)")

```

We can get a good assessment of how FFE and Jenkins calculations compare
just by comparing measured FFE and Jenkins computed carbon.

```{r}

compare_ffe_jenkins <- fvs_stand_measured_ffe |>
  mutate(FFE = Aboveground_Live) |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  select(STAND_ID, Year, FFE, `Stand Age`, `Forest Type`) |>
  full_join(fvs_stand_measured_jenkins, by = join_by(STAND_ID, Year)) |>
  mutate(Jenkins = Aboveground_Live) |>
  left_join(stand_first_measured, by = join_by(STAND_ID)) |>
  select(STAND_ID, Year, FFE, Jenkins, `Stand Age`, `Forest Type`)

ggplot(
    data = compare_ffe_jenkins
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = Jenkins, color = Year, shape = `Forest Type`)
  ) +
  ggtitle("FVS Carbon Calculation\nFFE vs. Jenkins Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

From this, it appears the Jenkins over-projects compared to FFE.

Same plot, but colored by stand age:

```{r}
ggplot(
    data = compare_ffe_jenkins
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = Jenkins, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("FVS Carbon Calculation\nFFE vs. Jenkins Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

### FIA Spatial

FIA has carbon in the tree table, in CARBON_AG; we can sum this per
stand and compute a per-hectare value.

Method:

-   Determine how much of each plot we will use the inventory

    -  Remove plots, subplots, and conditions that are not inventoried
       forest
       
    -  We are only including plots that are fully forested, with a single
       condition, so this is not relevant here.

-   Determine scale factor for each plot

    -  A full plot is 4 subplots
       of 24\^2\*pi square feet each; scale this according to how much
       we're using

-   scale the carbon in each tree for the area we're using to get carbon
    per acre

-   sum across all trees

Fetch all the trees for the grow-only plots:

```{r}
fia = DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')
fia_tree = tbl(fia, 'TREE')

# Create a dataframe containing the TREE rows for all inventories of
# all conditions for all subplots for all plots used in NK.
matching_tree <- fia_tree |>
  filter(INVYR >= 1999) |>
  # Only live trees
  filter(STATUSCD == 1) |>
  semi_join(
    stand_grow_only,
    # Fetch all conditions, all subplots
    by = join_by(STATECD, COUNTYCD, PLOT),
    copy = TRUE
  ) |>
  collect()

DBI::dbDisconnect(fia)

remove(fia, fia_tree)
```

For each tree, convert its carbon to carbon per acre, then to carbon per hectare.
To convert CARBON_AG, which is pounds, to carbon per acre, we observe that a
plot is 1/6 acre; replicating the plot 6 times gets us a full acre, so we
multiply the carbon in each tree by 6 to get lbs/acre of carbon.

```{r}
stand_measured_spatial <- matching_tree |>
  select(STATECD, COUNTYCD, PLOT, SUBP, CONDID, TREE, INVYR, CARBON_AG) |>
  mutate(CARBON_PER_ACRE = 6 * CARBON_AG) |>
  mutate(CARBON_METRIC = conv_multiunit(CARBON_PER_ACRE, "lbs / acre", "Mg / hectare")) |>
  group_by(STATECD, COUNTYCD, PLOT, INVYR) |>
  summarize(CARBON_METRIC = sum(CARBON_METRIC), .groups = "keep") |>
  ungroup() |>
  # Join to get STAND_IDs
  left_join(
    stand_id_crosswalk,
    by = join_by(STATECD, COUNTYCD, PLOT)
  ) |>
  # Join to get MEASYEAR, FORTYPE, and STDAGE
  left_join(
    matching_plot |> select(STATECD, COUNTYCD, PLOT, INVYR, MEASYEAR),
    by=join_by(STATECD, COUNTYCD, PLOT, INVYR)
  )

knitr::kable(stand_measured_spatial)
```

```{r}
compare_ffe_spatial <- fvs_stand_measured_ffe |>
  mutate(FFE = Aboveground_Live) |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  select(STAND_ID, Year, FFE, `Stand Age`, `Forest Type`) |>
  full_join(stand_measured_spatial, by = join_by(STAND_ID, Year == MEASYEAR)) |>
  mutate(Spatial = CARBON_METRIC) |>
  select(STAND_ID, Year, FFE, Spatial, `Stand Age`, `Forest Type`)

ggplot(
    data = compare_ffe_spatial
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = Spatial, color = Year, shape = `Forest Type`)
  ) +
  ggtitle("Carbon Calculation\nFFE vs. Spatial Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

Same, but colored by stand age

```{r}
ggplot(
    data = compare_ffe_spatial
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = Spatial, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("Carbon Calculation\nFFE vs. Spatial Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

### FIA TPA

FIA provides TREE.TPA_UNADJ which is the expansion factor to use for an individual
tree to get its carbon per acre contribution. CARBON_AG is pounds per tree,
so CARBON_AG * TPA_UNADJ is pounds per acre.

We have filtered to fully forested plots with a single condition, so the method
does not need to accommodate partially forested plots, or varying conditions.

Method:

- For each tree, compute CARBON_AG * TPA_UNADJ, then convert to Mg/ha.

- Sum for all trees in the plot


```{r}
stand_measured_tpa <- matching_tree |>
  select(STATECD, COUNTYCD, PLOT, SUBP, CONDID, TREE, INVYR, CARBON_AG, TPA_UNADJ) |>
  mutate(CARBON_PER_ACRE = CARBON_AG * TPA_UNADJ) |>
  mutate(CARBON_METRIC = conv_multiunit(CARBON_PER_ACRE, "lbs / acre", "Mg / hectare")) |>
  group_by(STATECD, COUNTYCD, PLOT, INVYR) |>
  summarize(CARBON_METRIC = sum(CARBON_METRIC), .groups = "keep") |>
  ungroup() |>
  # Join to get STAND_IDs
  left_join(
    stand_id_crosswalk,
    by = join_by(STATECD, COUNTYCD, PLOT)
  ) |>
  # Join to get MEASYEAR, FORTYPE, and STDAGE
  left_join(
    matching_plot |> select(STATECD, COUNTYCD, PLOT, INVYR, MEASYEAR),
    by=join_by(STATECD, COUNTYCD, PLOT, INVYR)
  )

knitr::kable(stand_measured_spatial)
```


```{r}
compare_ffe_tpa <- fvs_stand_measured_ffe |>
  mutate(FFE = Aboveground_Live) |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  select(STAND_ID, Year, FFE, `Stand Age`, `Forest Type`) |>
  full_join(stand_measured_tpa, by = join_by(STAND_ID, Year == MEASYEAR)) |>
  mutate(TPA = CARBON_METRIC) |>
  select(STAND_ID, Year, FFE, TPA, `Stand Age`, `Forest Type`)

ggplot(
    data = compare_ffe_tpa
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = TPA, color = Year, shape = `Forest Type`)
  ) +
  ggtitle("Carbon Calculation\nFFE vs. TPA Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

Same, but color by stand age

```{r}
ggplot(
    data = compare_ffe_tpa
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = FFE, y = TPA, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("Carbon Calculation\nFFE vs. TPA Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

These look very similar to the spatial approach; plot them against each other
to directly compare spatial vs. tpa

```{r}
compare_tpa_spatial <- fvs_stand_grow_only |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  select(STAND_ID, MEASYEAR, `Stand Age`, `Forest Type`) |>
  left_join(stand_measured_tpa, by = join_by(STAND_ID, MEASYEAR)) |>
  rename(TPA = CARBON_METRIC) |>
  left_join(stand_measured_spatial, by = join_by(STAND_ID, MEASYEAR)) |>
  rename(Spatial = CARBON_METRIC) |>
  rename(Year = MEASYEAR) |>
  select(STAND_ID, Year, Spatial, TPA, `Stand Age`, `Forest Type`)

ggplot(
    data = compare_tpa_spatial
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = Spatial, y = TPA, color = Year, shape = `Forest Type`)
  ) +
  ggtitle("FVS Carbon Calculation\nSpatial vs. TPA Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

Same, but color by stand age

```{r}
ggplot(
    data = compare_tpa_spatial
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = Spatial, y = TPA, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("Carbon Calculation\nSpatial vs. TPA Carbon (Mg/ha)") +
  coord_cartesian(xlim = c(0, 200), ylim = c(0, 200)) +
  geom_abline(intercept = 0, slope = 1)
```

## Basal Area

From this, we can see that FVS FFE over-projects carbon storage, and FVS Jenkins
over-projects even more. We can compare the measured vs. projected basal area
to get some idea whether the over-projection is part of the growth model or
a characteristic of the carbon calculation.

### Measured Basal Area

FIA provides basal area in COND.BALIVE, which is, per the FIADB User Guide:

> Basal area per acre of live trees. Basal area in square feet per acre of all live trees 1.0
inch d.b.h./d.r.c. sampled in the condition. Populated for forest conditions.

```{r}
stand_measured_balive <- matching_cond |>
  semi_join(stand_grow_only, by = join_by(STATECD, COUNTYCD, PLOT)) |>
  select(STATECD, COUNTYCD, PLOT, MEASYEAR, BALIVE) |>
  # BALIVE is square feet per acre; convert to square meters per hectare
  mutate(BALIVE_METRIC = conv_multiunit(BALIVE, "ft2 / acre", "m2 / hectare")) |>
  left_join(stand_id_crosswalk, by = join_by(STATECD, COUNTYCD, PLOT))

knitr::kable(stand_measured_balive)
```

FVS provides basal area in FVS_Summary2_East.BA, which is, per Essential FVS
section 4.3.1.4.1:

> basal area per acre after thinning

which we presume to be square feet per acre for standing live trees only.

```{r}
FVS_Summary2_East <- NULL

for (measyear in measyears)
  fname <- paste0("../03_outputs/Grow_", sprintf("%04d", measyear), "_FVS_Summary2_East.csv")
  FVS_Summary2_East <- bind_rows(
      FVS_Summary2_East,
      read_csv(fname, col_types = cols(StandID = col_character()))
    )

fvs_stand_measured_ba <- FVS_Summary2_East |>
  mutate(STAND_ID = paste0(
     substring(StandID, 1, 2),
     substring(StandID, 5, 12)
  )) |>
  inner_join(
    fvs_stand_grow_only |> select(STAND_ID, MEASYEAR, STDAGE, FORTYPE),
    by = join_by(STAND_ID, Year == MEASYEAR)
  ) |>
  mutate(BA_METRIC = conv_multiunit(BA, "ft2 / acre", "m2 / hectare")) |>
  select(STAND_ID, Year, BA, BA_METRIC, STDAGE, FORTYPE)

remove(FVS_Summary2_East, fname, measyear)

knitr::kable(fvs_stand_measured_ba)
```

```{r}
compare_fvs_ba_fia_balive <- stand_measured_balive |>
  rename(Year = MEASYEAR) |>
  rename(`FIA BALIVE` = BALIVE_METRIC) |>
  full_join(fvs_stand_measured_ba, by = join_by(STAND_ID, Year)) |>
  rename(`FVS BA` = BA_METRIC) |>
  rename(`Forest Type` = FORTYPE) |>
  rename(`Stand Age` = STDAGE) |>
  select(STAND_ID, Year, `FIA BALIVE`, `FVS BA`, `Forest Type`, `Stand Age`)

ggplot(
    data = compare_fvs_ba_fia_balive
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = `FIA BALIVE`, y = `FVS BA`, color = `Stand Age`, shape = `Forest Type`)
  ) +
  ggtitle("Measured Basal Area\nFIA BALIVE vs. FVS BA (m2 / ha)") +
  coord_cartesian(xlim = c(0, 60), ylim = c(0, 60)) +
  geom_abline(intercept = 0, slope = 1)
```

This shows that FVS and FIA approach to measuring basal area often aligns, but
there are cases when they do not align.

> TODO nik: what are the cases where these don't align? Looks like 6 points.

### Projected Basal Area

FVS also predicts basal area, and we can compare FVS projected with FVS measured.
This will give us insight just into the growth model of FVS independent of
carbon calculation.

```{r}
FVS_NoManagement_ByPlotYearly_Summary2 <-
  read_csv("../03_outputs/NoManagement_ByPlotYearly_Summary2.csv", 
    col_types = cols(StandID = col_character()))

fvs_stand_projected_ba <- FVS_NoManagement_ByPlotYearly_Summary2 |>
  mutate(STAND_ID = paste0(
     substring(StandID, 1, 2),
     substring(StandID, 5, 12)
  )) |>
  mutate(BA_METRIC = conv_multiunit(BA, "ft2 / acre", "m2 / hectare")) |>
  select(STAND_ID, Year, BA, BA_METRIC)

remove(FVS_NoManagement_ByPlotYearly_Summary2)

knitr::kable(fvs_stand_projected_ba)
```

```{r}
compare_fvs_ba_measured_projected <- fvs_stand_measured_ba |>
  rename(Measured = BA_METRIC) |>
  inner_join(fvs_stand_projected_ba, by = join_by(STAND_ID, Year)) |>
  rename(Projected = BA_METRIC) |>
  rename(`Stand Age` = STDAGE) |>
  rename(`Forest Type` = FORTYPE) |>
  inner_join(stand_first_measured, by = join_by(STAND_ID)) |>
  mutate(`Projection Years` = Year - first_measure_year) |>
  select(STAND_ID, Year, `Projection Years`, Measured, Projected, `Stand Age`, `Forest Type`)

ggplot(
    data = compare_fvs_ba_measured_projected
  ) +
  geom_point(
    size = 2,
    mapping = aes(x = Measured, y = Projected, color = `Projection Years`, shape = `Forest Type`)
  ) +
  ggtitle("Basal Area\nFVS Measured vs. Projected (m2 / ha)") +
  coord_cartesian(xlim = c(0, 60), ylim = c(0, 60)) +
  geom_abline(intercept = 0, slope = 1)
```
