==> 02_rScripts/06_LandisNoManagement.Rmd <==
---
title: "Nunery Keeton Revisited"
author: "Nikolaus Bates-Haus"
output:
  pdf_document: default
  html_document: default
editor_options:
  markdown:
    wrap: 72
---

```{r setup, include=FALSE}
# Pre-load libraries that issue warnings on startup so these
# warnings don't corrupt the text later.
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(reshape2)) # for melt()
suppressPackageStartupMessages(library(RSQLite))
suppressPackageStartupMessages(library(htmltools))
suppressPackageStartupMessages(library(rmarkdown))
suppressPackageStartupMessages(library(dbplyr))
suppressPackageStartupMessages(library(measurements))
suppressPackageStartupMessages(library(sf))
```

Load functions

```{r source-functions}
source('../R/functions.R')
```

## Species

### LandisData Species

`LandisData Species` wants a table with per-species metrics such as age
of maturity, etc.

The table currently in use says: \>\>NE LANDIS-II species data developed
from Thompson et al 2011, and if not present Duveneck et al. 2014.
\>\>Tamarack parameters develeped from Silvics manual and FEIS data
base. \>\>Matthew Duveneck 3/6/2014

## Ecoregions

### LandisData Ecoregions

ecoregions.txt is a table with one row per ecoregion in the map.
We only need a single ecoregion, so a single row.
We might include a row for MapCode 0 to include inactive pixels too.

| Attribute      | Value                 |
|----------------|-----------------------|
| Active         | yes                   |
| MapCode        | 1                     |
| ecoregion_name | ECOSUBCD, e.g. 211A   |
| description    | ECOSUBCD, e.g. "211A" |

ecoregions.img is a raster input with a single pixel in the image.

From the User Guide, section 3.2 Raster Input Maps:
> For a list of currently available formats, visit
> http://www.gdal.org/formats_list.html

```
% gdalinfo eco.img
Driver: HFA/Erdas Imagine Images (.img)
```

looks like it is format https://gdal.org/en/latest/drivers/raster/hfa.html#raster-hfa

No particular reason to change, other than not projecting is better than
projecting.

Projection?
- In eco.img: NAD83 / Conus Albers, EPSG:5070 
- Proposed: WGS 84, EPSG:4326
- USFS R9 Albers, ESRI:102042
- NAD83 / UTM zone 17N / 18N / 19N, EPSG 26917 26918 26919
- NAD83(2011) / UTM zone 17N / 18N / 19N, EPSG 6346 6347 6348
- WGS 84 UTM zone 17N / 18N / 19N, EPSG 32617 32618 32619

#### Creating Pixels

For each plot, create a 100x100 meter polygon centered at the location
of the plot.

Also also, from the sf docs:

> Why should we use OGC:CRS84 instead of EPSG:4326?

> EPSG:4326 formally defines coordinate axes to be in the order
> latitude-longitude, but practically all data sources and software
> environments use longitude-latitude axis order. OGC:CRS84 is
> equivalent to EPSG:4326 except that it defines coordinate axis order
> longitude-latitude, removing this ambiguity so to speak. See also
> st_axis_order()

```{r}
boston_plot_polygon <- hectare_at(42.360278, -71.057778)
st_area(boston_plot_polygon)
```

This gives the pixel position and shape, but not the value.
It looks like for the eco layer, intensity is in 'Band 1', or perhaps
'Layer_1'.
Layer Litter0 looks the same - "Band 1: Layer_1".

The raster says:
> Band 1 Block=64x64 Type=Int16, ColorInterp=Undefined
> Description = Layer_1
> Min=1.000 Max=1.000 
> Minimum=1.000, Maximum=1.000, Mean=1.000, StdDev=0.000
> NoData Value=-32768

So https://r-spatial.github.io/stars/articles/stars4.html has this example:
```
m = matrix(1:20, nrow = 5, ncol = 4)
dim(m) = c(x = 5, y = 4) # named dim
(s = st_as_stars(m))
```

and later
```
x = c(0, 0.5, 1, 2, 4, 5)  # 6 numbers: boundaries!
y = c(0.3, 0.5, 1, 2, 2.2) # 5 numbers: boundaries!
(r = st_as_stars(list(m = m), dimensions = st_dimensions(x = x, y = y)))
```

and from st_rasterize:
```
(x = st_rasterize(nc, st_as_stars(st_bbox(nc), nx = 100, ny = 50, values = NA_real_)))
```

OK, this might actually be right:
- Put the plot on the map as an st_point
- Convert to an appropriate CRS
- Put a circle around it using st_buffer https://r-spatial.github.io/sf/reference/geos_unary.html
- Rasterize the bounding box

This looks good:
```{r}
pixel <- c(42.36028, -71.05778) |>   # Raw lat / lon coordinates
  st_point() |>                      # Wrap in Point object
  st_sfc() |>                        # Wrap in Geometry Column
  st_sf(crs = 'EPSG:4326') |>        # Wrap in Simple Feature w/ lat/lon CRS
  st_transform(crs = 'EPSG:5070') |> # Transform to flat projection
  st_buffer(dist = 50) |>            # 50m buffer -> 100m diameter circle
  st_bbox() |>                       # bounding box -> 100m wide square
  st_as_stars(nx = 1, ny = 1, values = 9) # 1-by-1 Stars grid, assign value '9'
```

Writing this to gpkg is correct, except:
> Band 1 Block=256x256 Type=Float32, ColorInterp=Undefined
> Description = Height

We want Type to be byte, and Description to be Level_1
(or maybe 'ecoregion' - does LANDIS care?)

## PnET-Succession

### LandisData "PnET-Succession"

Timestep etc. need to align with the scenario.

Latitude is ceil(latitude) of the pixel.

### LandisData PNEToutputsites

Site1 1 1 << Single Cell

### LandisData "Initial Communities"

#### InitialCommunities.txt

Only one community, created to match the FIA plot.

Example:
```
MapCode 1 << FIA PLT_CN 73074175010538 << FIA Forest Type:  505 Northern red oak
fraxamer 50 (1561) 100 (1715)
pinustro 100 (990)
querrubr 50 (2454) 100 (9592)
tsugcana 50 (1699)
```

#### InitialCommunities.img

Single pixel, must match ecoregion, 

### What species occur where?

What species are in which ecosubregion?

```{r}

# TREE.SPCD - species code
# FIADB_REFERENCE.REF_SPECIES.SPCD - species code
# FIADB_REFERENCE.REF_SPECIES.SCIENTIFIC_NAME / GENUS / SPECIES
# TREE.PLT_CN - Plot Control Number
# PLOT.CN - Plot Control Number
# PLOT.ECOSUBCD

fia_ref <- DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/FIADB_REFERENCE.db')
fia_ref_species <- tbl(fia_ref, 'REF_SPECIES') |> 
  select(
    SPCD, SCIENTIFIC_NAME, GENUS, SPECIES, COMMON_NAME
  ) |> 
  collect()
dbDisconnect(fia_ref)


fia <- DBI::dbConnect(RSQLite::SQLite(), '../00_rawData/SQLite_FIADB_ENTIRE.db')

fia_tree <- tbl(fia, 'TREE') |> 
  distinct(PLT_CN, SPCD)

fia_plot <- tbl(fia, 'PLOT') |>
  select(CN, SRV_CN, ECOSUBCD) |>
  rename(PLT_CN = CN)

fia_survey <- tbl(fia, 'SURVEY') |>
  select(CN, RSCD) |>
  rename(SRV_CN = CN)

species_by_ecosubcd <- fia_tree |>
  left_join(fia_plot, by = join_by(PLT_CN)) |>
  left_join(fia_survey, by = join_by(SRV_CN)) |>
  northeastern_plots_filter() |>
  distinct(ECOSUBCD, SPCD) |>
  collect() |> 
  left_join(fia_ref_species, by = join_by(SPCD))

dbDisconnect(fia)
remove(fia, fia_tree, fia_plot, fia_survey, fia_ref, fia_ref_species)
```
